requirements.txt
ipython
asgiref==3.8.1
Django==5.1.3
django-extensions
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
pokebase==1.3.0
djangorestframework
pytest
pytest-mock
